Convert dbplyr table to sf Object

Collects a lazy dbplyr view and materializes it as an in-memory sf table.

Usage

collect_sf(tbl, geom_col = NULL, crs = 4326)

Arguments

tbl

A dbplyr view object containing a 'geometry' column.

geom_col

The name of the geometry column. Will auto-detect names matching 'geom'.

crs

The coordinate reference system to use for the geometries, specified by its EPSG code. The default is 4326 (WGS 84).

Value

An 'sf' object with the dataset converted to spatial features.

Examples

if (FALSE) { # interactive()

bbox <- c(xmin = -120.5, ymin = 35.5, xmax = -120.0, ymax = 36.0)
lazy_tbl <- open_curtain("building", bbox)
collect_sf(lazy_tbl)
}